pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nl2scene3d"
version = "0.1.0"
description = "Scene reorganization from random to ordered via Multimodal Language Models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "NL2Scene3D Team" }]
dependencies = [
    "google-generativeai>=0.8.0",
    "google-api-core>=2.15.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations
]
ignore = [
    "ANN101",  # missing type annotation for self (deprecated in newer ruff)
    "ANN102",  # missing type annotation for cls (deprecated in newer ruff)
    "ANN401",  # disallows Any type annotation
]

[tool.ruff.lint.isort]
known-first-party = ["nl2scene3d"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/nl2scene3d --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = ["src/nl2scene3d"]